Feat: Splitting pyattck_data and generation of external data
