+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyattck-data"
-version = "2.1.1"
+version = "3.0.0"
 description = "Pyattck Data"
 authors = ["Swimlane <info@swimlane.com>"]
 license = "MIT"
@@ -16,7 +16,6 @@ Changelog = "https://github.com/swimlane/pyattck-data/releases"
 
 [tool.poetry.dependencies]
 python = "^3.7"
-click = "^8.0.1"
 attrs = "^21.4.0"
 pydantic = "^1.9.1"
 requests = "^2.28.0"
@@ -53,9 +52,6 @@ Pygments = "^2.10.0"
 pyupgrade = "^2.29.1"
 furo = ">=2021.11.12"
 requests = "^2.27.1"
-
-[tool.poetry.scripts]
-pyattck-data-models = "pyattck_data.__main__:main"
 
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
