--- src/pyattck_data/__init__.py
@@ -1,2 +1,0 @@
-from .pyattckdata import PyattckData
-from .generatenistdata import GenerateNISTData
