+++ src/pyattck_data/malware.py
