--- src/pyattck_data/services/__init__.py
@@ -1,19 +1,0 @@
-from .adversaryemulation import AdversaryEmulation
-from .aptthreattracking import APTThreatTracking
-from .atomicredteam import AtomicRedTeam
-from .atomicthreatcoverage import AtomicThreatCoverage
-from .attckdatasources import AttckDatasources
-from .attckempire import AttckEmpire
-from .blueteamlabs import BlueTeamLabs
-from .c2matrix import C2Matrix
-from .elemental import ElementalAttack
-from .litmustest import LitmusTest
-from .macosattackdataset import MacOSAttackDataset
-from .malwarearchaeology import MalwareArchaeology
-from .newbeeattackdata import NewBeeAttackDataset
-from .nsmattck import NSMAttck
-from .osqueryattack import OsqueryAttack
-from .stockpile import MitreStockpile
-from .sysmonhunter import SysmonHunter
-from .threathuntingbook import ThreatHuntingBook
-from .threathuntingtables import ThreatHuntingTables
